requirements.txt
-i https://pypi.python.org/simple
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
click==7.0
flask-bcrypt==0.7.1
flask-cors==3.0.7
flask-login==0.4.1
flask-marshmallow==0.9.0
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==1.0.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10
markupsafe==1.1.0
marshmallow==2.18.1
psycopg2==2.7.7
pycparser==2.19
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
sqlalchemy==1.2.18
urllib3==1.24.1
werkzeug==0.14.1
wtforms==2.2.1
